Add the VisualNavigator protocol
