+++ r2-navigator-swift/Navigator.swift
@@ -10,19 +10,12 @@
 //
 
 import Foundation
-import UIKit
 import SafariServices
 import R2Shared
 
 
 public protocol Navigator {
     
-    /// Viewport view.
-    var view: UIView! { get }
-    
-    /// Current reading progression.
-    var readingProgression: ReadingProgression { get }
-
     /// Current position in the publication.
     /// Can be used to save a bookmark to the current position.
     var currentLocation: Locator? { get }
@@ -50,73 +43,40 @@ public protocol Navigator {
     /// - Returns: Whether the navigator is able to move to the previous content portion. The completion block is only called if true was returned.
     @discardableResult
     func goBackward(animated: Bool, completion: @escaping () -> Void) -> Bool
-    
-    /// Moves to the left content portion (eg. page) relative to the reading progression direction.
-    /// - Parameter completion: Called when the transition is completed.
-    /// - Returns: Whether the navigator is able to move to the previous content portion. The completion block is only called if true was returned.
-    @discardableResult
-    func goLeft(animated: Bool, completion: @escaping () -> Void) -> Bool
-    
-    /// Moves to the right content portion (eg. page) relative to the reading progression direction.
-    /// - Parameter completion: Called when the transition is completed.
-    /// - Returns: Whether the navigator is able to move to the previous content portion. The completion block is only called if true was returned.
-    @discardableResult
-    func goRight(animated: Bool, completion: @escaping () -> Void) -> Bool
-    
+
 }
 
-/// Adds default values for the methods' parameters.
 public extension Navigator {
     
+    /// Adds default values for the parameters.
     @discardableResult
     func go(to locator: Locator, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
         return go(to: locator, animated: animated, completion: completion)
     }
     
+    /// Adds default values for the parameters.
     @discardableResult
     func go(to link: Link, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
         return go(to: link, animated: animated, completion: completion)
     }
     
+    /// Adds default values for the parameters.
     @discardableResult
     func goForward(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
         return goForward(animated: animated, completion: completion)
     }
     
+    /// Adds default values for the parameters.
     @discardableResult
     func goBackward(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
         return goBackward(animated: animated, completion: completion)
     }
-    
-    @discardableResult
-    func goLeft(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        switch readingProgression {
-        case .ltr, .auto:
-            return goBackward(animated: animated, completion: completion)
-        case .rtl:
-            return goForward(animated: animated, completion: completion)
-        }
-    }
-    
-    @discardableResult
-    func goRight(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        switch readingProgression {
-        case .ltr, .auto:
-            return goForward(animated: animated, completion: completion)
-        case .rtl:
-            return goBackward(animated: animated, completion: completion)
-        }
-    }
-    
+
 }
 
 
 public protocol NavigatorDelegate: AnyObject {
-    
-    /// Called when the user tapped the publication, and it didn't trigger any internal action.
-    /// The point is relative to the navigator's view.
-    func navigator(_ navigator: Navigator, didTapAt point: CGPoint)
-    
+
     /// Called when the current position in the publication changed. You should save the locator here to restore the last read page.
     func navigator(_ navigator: Navigator, locationDidChange locator: Locator)
     
@@ -134,11 +94,7 @@ public extension NavigatorDelegate {
     func navigator(_ navigator: Navigator, presentExternalURL url: URL) {
         UIApplication.shared.openURL(url)
     }
-    
-    func navigator(_ navigator: Navigator, didTapAt point: CGPoint) {
-        // Optional
-    }
-    
+
 }
